move window title files to their own folder
