+++ src/electron/views/common/window-title/window-title.scss
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-@import '../../../common/styles/colors.scss';
-@import '../../../common/styles/fonts.scss';
+@import '../../../../common/styles/colors.scss';
+@import '../../../../common/styles/fonts.scss';
 @import './window-title-const.scss';
 
 .window-title {
