+++ src/electron/views/common/window-title/window-title.tsx
